requirements.txt
gradio
openai
python-dotenv
manim
pydantic
pydantic-ai
google-generativeai
aiohttp